transition predictions
